moving abstract class parsing to parse_class_statement
