+++ crates/rslint_parser/test_data/inline/err/abstract_class_in_js.rast
@@ -32,7 +32,7 @@ JsModule {
       5: R_CURLY@18..19 "}" [] []
   3: EOF@19..20 "" [Newline("\n")] []
 --
-error[SyntaxError]: `abstract classes` can only be declared in TypeScript files
+error[SyntaxError]: `abstract` classes can only be declared in TypeScript files
   ┌─ abstract_class_in_js.js:1:1
   │
 1 │ abstract class A {}
