+++ crates/rslint_parser/test_data/inline/err/typescript_class_member_body.rast
@@ -605,28 +605,28 @@ error[SyntaxError]: abstract members should not have a body
   ┌─ typescript_class_member_body.ts:9:28
   │
 9 │     abstract constructor() { }
-  │                            ^^^ unexpected
+  │                            ^^^
 
 --
 error[SyntaxError]: abstract members should not have a body
    ┌─ typescript_class_member_body.ts:10:30
    │
 10 │     abstract display(): void { console.log(this.name); }
-   │                              ^^^^^^^^^^^^^^^^^^^^^^^^^^^ unexpected
+   │                              ^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 --
 error[SyntaxError]: abstract members should not have a body
    ┌─ typescript_class_member_body.ts:11:28
    │
 11 │     abstract get my_name() { return this.name; }
-   │                            ^^^^^^^^^^^^^^^^^^^^^ unexpected
+   │                            ^^^^^^^^^^^^^^^^^^^^^
 
 --
 error[SyntaxError]: abstract members should not have a body
    ┌─ typescript_class_member_body.ts:12:32
    │
 12 │     abstract set my_name(name) { this.name = name; }
-   │                                ^^^^^^^^^^^^^^^^^^^^^ unexpected
+   │                                ^^^^^^^^^^^^^^^^^^^^^
 
 --
 error[SyntaxError]: members with private name cannot be abstract
@@ -640,7 +640,7 @@ error[SyntaxError]: abstract members should not have a body
    ┌─ typescript_class_member_body.ts:13:30
    │
 13 │     abstract #private_name() { }
-   │                              ^^^ unexpected
+   │                              ^^^
 
 --
 class AbstractMembers {
